backend/requirements.txt
fastapi
uvicorn
requests
python-dotenv
sqlalchemy
pymysql
cryptography
bcrypt
pytest
pytest-asyncio
httpx
pyjwt

